+++ Extension/YouTubeDL.RoughDraft.Extension.ps1
@@ -1,5 +1,14 @@
+<#
+.SYNOPSIS
+    Youtube Downloader
+.DESCRIPTION
+    Extends Get-Media to enable the downloading of videos from YouTube and other sources, using YouTubeDL
+.LINK
+    http://ytdl-org.github.io/youtube-dl/
+#>
 [Management.Automation.Cmdlet("Get","Media")]
-[ComponentModel.InheritanceLevel("NotInherited")]
+# that is not inherited (this is the default).
+[ComponentModel.Inheritance("NotInherited")]
 param(
 [Parameter(Mandatory)]
 [Alias('YouTubeURI')]